+++ spec/nylas/client_spec.rb
@@ -47,6 +47,7 @@ describe Nylas::Client do
         expect(nylas.drafts).to be_a(Nylas::Drafts)
         expect(nylas.events).to be_a(Nylas::Events)
         expect(nylas.grants).to be_a(Nylas::Grants)
+        expect(nylas.lists).to be_a(Nylas::Lists)
         expect(nylas.folders).to be_a(Nylas::Folders)
         expect(nylas.messages).to be_a(Nylas::Messages)
         expect(nylas.threads).to be_a(Nylas::Threads)